feat: 完成项目全量初始化与功能迭代
整体提交包含：
1.  新增移动端Android/iOS原生工程配置、图标资源与基础框架
2.  新增后端管理后台、多语言国际化与权限验证
3.  新增情感分析、关系管理、支付推送等业务模块
4.  更新依赖库与数据库结构，补充完整测试用例
5.  新增本地调试脚本与gitignore配置

+++ services/backend/alembic/versions/001_initial_schema.py
@@ -48,7 +48,7 @@ def upgrade() -> None:
         sa.Column("source", sa.String(200), nullable=False),
         sa.Column("description", sa.Text, nullable=False),
         sa.Column("avatar_url", sa.String(500), server_default=""),
-        sa.Column("color", sa.Integer, server_default="0"),
+        sa.Column("color", sa.BigInteger, server_default="0"),
         sa.Column("personality", JSONB, nullable=False),
         sa.Column("system_prompt", sa.Text, nullable=False),
         sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.func.now()),
